Flesh out the Program run code to cover all cases
Previously the runProgramInvocation impl only covered the common special
cases that we currently use. In particular it did not support env vars
or changing the workind directory.

Also change the ProgramInvocation { progInvokeEnv } to allow unsetting
env vars, not just setting them to new values.

This stuff would be better if we could use createProcess for all cases
but that doesn't yet support running programs in the foreground with
proper ctl-C handling (only rawSystem currently supports that).

Conflicts:
	Cabal/Distribution/Simple/Test.hs
	Cabal/Distribution/Simple/Utils.hs